Refactored to use a CSV as symbol DB, refactored tensorflowstoraged, out of metadata module
