+++ src/generator/cad_converter.py
@@ -13,14 +13,13 @@ logger = logging.getLogger()
 
 
 def find_symbol_file(symbol_name: str) -> str:
-    path = check_output(
-        f"find {DATA_PATH}/symbol_libraries/official -name {symbol_name}.dwg".split(" ")
+    command = f"find {DATA_PATH}/symbol_libraries/official -name {symbol_name}.dwg".split(
+        " "
     )
+    path = check_output(command)
     if len(path) > 2:
         return path.strip().decode()
-    else:
-        logger.debug(f"Cannot find file for: {symbol_name}")
-    raise Exception("File not found")
+    raise Exception(f"File not found {command}")
 
 
 def collect_dwg_file(file_path):
