+++ src/generator/launch_set_up_symbol_library.py
@@ -4,7 +4,7 @@
 import argparse
 import logging
 
-from generator.cad_converter import find_symbol_file, collect_dwg_file, dxf_to_png
+from generator.cad_converter import find_symbol_file, collect_dwg_file
 from generator.ccf_reader import parse_ccf_file
 from generator.metadata import SymbolStorage
 
@@ -21,7 +21,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     if args.ccf_filename:
-        symbols = parse_ccf_file(args.ccf_filename[0])
+        symbols = parse_ccf_file(args.ccf_filename)
         symbols_with_path = []
         for symbol in symbols:
             symbol_name = symbol[0]
